refactor new project view, move draw logic into draw-control
