+++ app/components/draw-control.js
@@ -1,7 +1,18 @@
 import Component from '@ember/component';
 import { action } from '@ember-decorators/object';
 import { argument } from '@ember-decorators/argument';
+import MapboxDraw from 'mapbox-gl-draw';
+import turfUnion from 'npm:@turf/union';
+import turfBuffer from 'npm:@turf/buffer';
 
+const draw = new MapboxDraw({
+  displayControlsDefault: false,
+  controls: {
+    polygon: true,
+    trash: true,
+  },
+  // styles: drawStyles, TODO modify default draw styles
+});
 
 export default class DrawControlController extends Component {
   @argument
@@ -29,25 +40,87 @@ export default class DrawControlController extends Component {
   finishLotSelection
 
   @argument
+  selectedLots
+
+  @argument
+  lotSelectionMode
+
+  @argument
   setProjectGeometry
 
-  @action
-  toggleGeometryEditing(type) {
-    this.get('toggleGeometryEditing')(type);
+  @argument
+  mapInstance
+
+  @argument
+  model
+
+  @action toggleGeometryEditing(type) {
+    this.set('projectGeometryMode', type);
+
+    const projectGeometryMode = this.get('projectGeometryMode');
+
+    const map = this.get('mapInstance');
+    if (projectGeometryMode) {
+      map.addControl(draw, 'top-right');
+      draw.changeMode('draw_polygon');
+    } else {
+      draw.trash();
+      map.removeControl(draw);
+    }
   }
 
   @action
   startLotSelection() {
-    this.get('startLotSelection')();
+    this.set('lotSelectionMode', true);
+    draw.trash();
+    draw.changeMode('simple_select');
   }
 
   @action
   finishLotSelection() {
-    this.get('finishLotSelection')();
+    // runs when user is done selecting lots
+    // adds a small buffer to all lots to ensure the union will be contiguous
+    // unions all lots together into one feature
+    // TODO simplify the resulting union to get rid of curved corners with many vertices
+    this.set('lotSelectionMode', false);
+
+    const selectedLots = this.get('selectedLots');
+    const bufferkm = 0.00008;
+
+    let union = turfBuffer(selectedLots.features[0].geometry, bufferkm);
+
+    if (selectedLots.features.length > 1) {
+      for (let i = 1; i < selectedLots.features.length; i += 1) {
+        const bufferedGeometry = turfBuffer(selectedLots.features[i].geometry, bufferkm);
+
+        union = turfUnion.default(union, bufferedGeometry);
+      }
+    }
+
+    // set the drawn geom as an editable mapbox-gl-draw geom
+    draw.set({
+      type: 'FeatureCollection',
+      features: [union],
+    });
+    // clear selected features
+    this.set('selectedLots.features', []);
   }
 
   @action
   setProjectGeometry() {
-    this.get('setProjectGeometry')();
+    const FeatureCollection = draw.getAll();
+
+    // delete the drawn geometry
+    draw.deleteAll();
+
+    const { geometry } = FeatureCollection.features[0];
+    const projectGeometryMode = this.get('projectGeometryMode');
+
+    // set geometry depending on mode
+    const model = this.get('model');
+    model.set(projectGeometryMode, geometry);
+
+    // breakdown the draw tools
+    this.toggleGeometryEditing(null);
   }
 }
