+++ app/helpers/mapbox-geojson-source.js
@@ -1,0 +1,10 @@
+import { helper } from '@ember/component/helper';
+
+export function mapboxGeojsonSource([data]) {
+  return {
+    type: 'geojson',
+    data,
+  };
+}
+
+export default helper(mapboxGeojsonSource);
