+++ app/templates/projects/new.hbs
@@ -17,38 +17,37 @@
 
   <label class="text-small">Project Geometries</label>
 
-  {{draw-control
-    mode="developmentSite"
-    modeDisplayName="Development Site"
-    projectGeometryMode=projectGeometryMode
-    lotSelectionMode=lotSelectionMode
-    toggleGeometryEditing=(action "toggleGeometryEditing")
-    startLotSelection=(action "startLotSelection")
-    finishLotSelection=(action "finishLotSelection")
-    setProjectGeometry=(action "setProjectGeometry")
-  }}
+  {{#if mapInstance}}
+    {{draw-control
+      model=model
+      mapInstance=mapInstance
+      mode="developmentSite"
+      modeDisplayName="Development Site"
+      projectGeometryMode=projectGeometryMode
+      selectedLots=selectedLots
+      lotSelectionMode=lotSelectionMode
+    }}
 
-  {{draw-control
-    mode="projectArea"
-    modeDisplayName="Project Area"
-    projectGeometryMode=projectGeometryMode
-    lotSelectionMode=lotSelectionMode
-    toggleGeometryEditing=(action "toggleGeometryEditing")
-    startLotSelection=(action "startLotSelection")
-    finishLotSelection=(action "finishLotSelection")
-    setProjectGeometry=(action "setProjectGeometry")
-  }}
+    {{draw-control
+      model=model
+      mapInstance=mapInstance
+      mode="projectArea"
+      modeDisplayName="Project Area"
+      projectGeometryMode=projectGeometryMode
+      selectedLots=selectedLots
+      lotSelectionMode=lotSelectionMode
+    }}
 
-  {{draw-control
-    mode="rezoningArea"
-    modeDisplayName="Rezoning Area"
-    projectGeometryMode=projectGeometryMode
-    lotSelectionMode=lotSelectionMode
-    toggleGeometryEditing=(action "toggleGeometryEditing")
-    startLotSelection=(action "startLotSelection")
-    finishLotSelection=(action "finishLotSelection")
-    setProjectGeometry=(action "setProjectGeometry")
-  }}
+    {{draw-control
+      model=model
+      mapInstance=mapInstance
+      mode="rezoningArea"
+      modeDisplayName="Rezoning Area"
+      projectGeometryMode=projectGeometryMode
+      selectedLots=selectedLots
+      lotSelectionMode=lotSelectionMode
+    }}
+  {{/if}}
 
   <button type="button" class="button expanded project-save-button" disabled={{not model.isValid}} onclick={{action 'save' model}}>Create Project</button>
 
@@ -72,7 +71,7 @@
 
         {{!-- Ad-hoc source and layer for developmentSite --}}
         {{#if model.developmentSite}}
-          {{#map.source options=developmentSiteSource as |source|}}
+          {{#map.source options=(mapbox-geojson-source model.developmentSite) as |source|}}
             {{source.layer
               layer=developmentSiteLayer
               before='boundary_country'
@@ -82,7 +81,7 @@
 
         {{!-- Ad-hoc source and layer for projectArea --}}
         {{#if model.projectArea}}
-          {{#map.source options=projectAreaSource as |source|}}
+          {{#map.source options=(mapbox-geojson-source model.projectArea) as |source|}}
             {{source.layer
               layer=projectAreaLayer
               before='boundary_country'
@@ -92,7 +91,7 @@
 
         {{!-- Ad-hoc source and layer for projectArea --}}
         {{#if model.rezoningArea}}
-          {{#map.source options=rezoningAreaSource as |source|}}
+          {{#map.source options=(mapbox-geojson-source model.rezoningArea) as |source|}}
             {{source.layer
               layer=rezoningAreaLayer
               before='boundary_country'
