+++ tests/integration/helpers/mapbox-geojson-source-test.js
@@ -1,0 +1,17 @@
+import { module, test } from 'qunit';
+import { setupRenderingTest } from 'ember-qunit';
+import { render } from '@ember/test-helpers';
+import hbs from 'htmlbars-inline-precompile';
+
+module('Integration | Helper | mapbox-geojson-source', function(hooks) {
+  setupRenderingTest(hooks);
+
+  // Replace this with your real tests.
+  test('it renders', async function(assert) {
+    this.set('inputValue', '1234');
+
+    await render(hbs`{{mapbox-geojson-source inputValue}}`);
+
+    assert.equal(this.element.textContent.trim(), '1234');
+  });
+});
